Phase 0: extract shared substrate into utils.py
Move the numerical backend config (config(), dynamic globals, rngkey, and the
optional matfree cgsolve/make_logdet_estimator block) and the kernel-primitive
markers + indexed Sherman-Morrison helpers out of matrix.py / kernel_helpers.py
into a new path-neutral utils.py.

- matrix.py re-exports these from utils so legacy matrix.<name> access keeps
  working for signals.py / likelihood.py.
- metamath.py and likelihood_metamath.py read markers and config aliases from
  utils instead of kernel_helpers.
- kernel_helpers.py deleted; its __getattr__ matrix-proxy is no longer needed
  since config now lives in the neutral module.

This breaks the kernel_helpers -> matrix back-reference, giving both kernel
paths a common substrate -- a prerequisite for eventually deleting matrix.py.

Full suite green (43 parity + 29 matrix-path tests).
